wsRecv/requirement.txt
websocket-client>=0.56.0
pika>=1.1.0
